add payment intent route to auth
