+++ backend/routes/auth.js
@@ -4,7 +4,7 @@ var express = require('express');
 var router = express.Router();
 var purchasesRouter = require('./auth/purchases');
 var testingRouter = require('./auth/testing');
-
+var paymentIntent = require('./auth/payment_intent');
 var serviceAccount = {
   "type": "service_account",
   "project_id": "iron-entropy-285002",
@@ -41,5 +41,6 @@ function checkAuth(req, res, next) {
 router.use('/', checkAuth);
 router.use('/purchases', purchasesRouter);
 router.use('/testing', testingRouter);
+router.use('/create-payment-intent', paymentIntent);
 
 module.exports = router;
